Fix folder deletion to be userId aware

+++ routes/folders.js
@@ -128,6 +128,7 @@ router.put('/:id', (req, res, next) => {
 /* ========== DELETE/REMOVE A SINGLE ITEM ========== */
 router.delete('/:id', (req, res, next) => {
   const { id } = req.params;
+  const { id: userId } = req.user;
 
   /** *** Never trust users - validate input **** */
   if (!mongoose.Types.ObjectId.isValid(id)) {
@@ -136,7 +137,7 @@ router.delete('/:id', (req, res, next) => {
     return next(err);
   }
 
-  Folder.findOneAndDelete({ _id: id })
+  Folder.findOneAndDelete({ _id: id, userId })
     .then((deleted) => {
       if (!deleted) {
         return;
